+++ src/net/CCommManager.h
@@ -70,9 +70,9 @@ public:
     virtual void AddReceiver(ReceiverRecord *aReceiver, Boolean delayed);
     virtual void RemoveReceiver(ReceiverRecord *aReceiver, Boolean delayed);
 
-    virtual OSErr SendPacket(short distribution, char command, char p1, short p2, long p3, short dataLen, Ptr dataPtr);
+    virtual OSErr SendPacket(short distribution, char command, char p1, int16_t p2, int32_t p3, int16_t dataLen, Ptr dataPtr);
     virtual OSErr
-    SendUrgentPacket(short distribution, char command, char p1, short p2, long p3, short dataLen, Ptr dataPtr);
+    SendUrgentPacket(short distribution, char command, char p1, int16_t p2, int32_t p3, int16_t dataLen, Ptr dataPtr);
     virtual void Dispose();
 
     virtual PacketInfo *GetPacket();
